move : inputchecer, randomUtils 패키지 이동
